rename multiple_canister_method_classes to multiple_canister_classes
