+++ examples/stable/test/end_to_end/candid_rpc/multiple_canister_classes/dfx.json
@@ -1,10 +1,10 @@
 {
     "canisters": {
-        "multiple_canister_method_classes": {
+        "multiple_canister_classes": {
             "type": "azle",
             "main": "src/index.ts",
             "declarations": {
-                "output": "test/dfx_generated/multiple_canister_method_classes",
+                "output": "test/dfx_generated/multiple_canister_classes",
                 "node_compatibility": true
             }
         }
