+++ examples/stable/test/end_to_end/candid_rpc/multiple_canister_classes/src/index.ts
