+++ examples/stable/test/end_to_end/candid_rpc/multiple_canister_classes/test/pretest.ts
@@ -1,0 +1,17 @@
+import { execSync } from 'child_process';
+
+function pretest(): void {
+    execSync(`dfx canister uninstall-code multiple_canister_classes || true`, {
+        stdio: 'inherit'
+    });
+
+    execSync(`dfx deploy multiple_canister_classes`, {
+        stdio: 'inherit'
+    });
+
+    execSync(`dfx generate multiple_canister_classes`, {
+        stdio: 'inherit'
+    });
+}
+
+pretest();
