--- examples/stable/test/end_to_end/candid_rpc/multiple_canister_method_classes/test/test.ts
@@ -1,17 +1,0 @@
-import { getCanisterId } from 'azle/_internal/dfx';
-import { runTests } from 'azle/_internal/test';
-
-import { createActor } from './dfx_generated/multiple_canister_method_classes';
-import { getTests } from './tests';
-
-const canister = createActor(
-    getCanisterId('multiple_canister_method_classes'),
-    {
-        agentOptions: {
-            host: 'http://127.0.0.1:4943',
-            shouldFetchRootKey: true
-        }
-    }
-);
-
-runTests(getTests(canister));
